Add module for Java 16 validation tests (#1165)
Records and Pattern Matching for instanceof are finalized in Java 16,
so no need to test them as preview features in earlier versions.
