+++ org.jacoco.core.test.validation.java16/pom.xml
@@ -20,12 +20,12 @@
     <relativePath>../org.jacoco.core.test.validation</relativePath>
   </parent>
 
-  <artifactId>org.jacoco.core.test.validation.java14</artifactId>
+  <artifactId>org.jacoco.core.test.validation.java16</artifactId>
 
-  <name>JaCoCo :: Test :: Core :: Validation Java 14</name>
+  <name>JaCoCo :: Test :: Core :: Validation Java 16</name>
 
   <properties>
-    <bytecode.version>14</bytecode.version>
+    <bytecode.version>16</bytecode.version>
   </properties>
 
   <dependencies>
