+++ org.jacoco.core.test.validation.java16/src/org/jacoco/core/test/validation/java16/RecordsTest.java
@@ -10,28 +10,18 @@
  *    Evgeny Mandrikov - initial API and implementation
  *
  *******************************************************************************/
-package org.jacoco.core.test.validation.java14;
+package org.jacoco.core.test.validation.java16;
 
-import org.jacoco.core.test.validation.Source.Line;
 import org.jacoco.core.test.validation.ValidationTestBase;
-import org.jacoco.core.test.validation.java14.targets.InstanceofTarget;
+import org.jacoco.core.test.validation.java16.targets.RecordsTarget;
 
 /**
- * Test of code coverage in {@link InstanceofTarget}.
+ * Test of code coverage for records.
  */
-public class InstanceofTest extends ValidationTestBase {
+public class RecordsTest extends ValidationTestBase {
 
-	public InstanceofTest() {
-		super(InstanceofTarget.class);
-	}
-
-	public void assertInstanceof(final Line line) {
-		if (isJDKCompiler && JAVA_VERSION.isBefore("15")) {
-			// https://bugs.openjdk.java.net/browse/JDK-8237528
-			assertFullyCovered(line, 1, 3);
-		} else {
-			assertFullyCovered(line, 0, 2);
-		}
+	public RecordsTest() {
+		super(RecordsTarget.class);
 	}
 
 }
