+++ org.jacoco.core.test.validation.java16/src/org/jacoco/core/test/validation/java16/targets/InstanceofTarget.java
@@ -10,17 +10,18 @@
  *    Evgeny Mandrikov - initial API and implementation
  *
  *******************************************************************************/
-package org.jacoco.core.test.validation.java14.targets;
+package org.jacoco.core.test.validation.java16.targets;
 
 import static org.jacoco.core.test.validation.targets.Stubs.nop;
 
 /**
- * This target exercises pattern matching for instanceof (JEP 305).
+ * This target exercises pattern matching for instanceof
+ * (<a href="https://openjdk.java.net/jeps/394">JEP 394</a>).
  */
 public class InstanceofTarget {
 
 	private static void ifInstanceof(Object e) {
-		if (e instanceof String s) { // assertInstanceof()
+		if (e instanceof String s) { // assertFullyCovered(0, 2)
 			nop(s);
 		}
 	}
