+++ org.jacoco.doc/docroot/doc/build.html
@@ -97,6 +97,21 @@
   </thead>
   <tbody>
   <tr>
+    <td>org.jacoco.core.test.validation.java16</td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td></td>
+    <td>16</td>
+  </tr>
+  <tr>
     <td>org.jacoco.core.test.validation.java14</td>
     <td></td>
     <td></td>
